+++ pkg/resource/deploy/snapshot.go
@@ -24,6 +24,7 @@ import (
 	"github.com/pulumi/pulumi/pkg/resource"
 	"github.com/pulumi/pulumi/pkg/resource/deploy/providers"
 	"github.com/pulumi/pulumi/pkg/secrets"
+	"github.com/pulumi/pulumi/pkg/util/contract"
 	"github.com/pulumi/pulumi/pkg/workspace"
 )
 
@@ -64,6 +65,52 @@ func NewSnapshot(manifest Manifest, secretsManager secrets.Manager,
 		SecretsManager:    secretsManager,
 		Resources:         resources,
 		PendingOperations: ops,
+	}
+}
+
+// NormalizeURNReferences fixes up all URN references in a snapshot to use the new URNs instead of potentially-aliased
+// URNs.  This will affect resources that are "old", and which would be expected to be updated to refer to the new names
+// later in the deployment.  But until they are, we still want to ensure that any serialization of the snapshot uses URN
+// references which do not need to be indirected through any alias lookups, and which instead refer directly to the URN
+// of a resource in the resources map.
+//
+// Note: This method modifies the snapshot (and resource.States in the snapshot) in-place.
+func (snap *Snapshot) NormalizeURNReferences() {
+	if snap != nil {
+		aliased := make(map[resource.URN]resource.URN)
+		fixUrn := func(urn resource.URN) resource.URN {
+			if newUrn, has := aliased[urn]; has {
+				return newUrn
+			}
+			return urn
+		}
+		for _, state := range snap.Resources {
+			// Fix up any references to URNs
+			state.Parent = fixUrn(state.Parent)
+			for i, dependency := range state.Dependencies {
+				state.Dependencies[i] = fixUrn(dependency)
+			}
+			for k, deps := range state.PropertyDependencies {
+				for i, dep := range deps {
+					state.PropertyDependencies[k][i] = fixUrn(dep)
+				}
+			}
+			if state.Provider != "" {
+				ref, err := providers.ParseReference(state.Provider)
+				contract.AssertNoError(err)
+				ref, err = providers.NewReference(fixUrn(ref.URN()), ref.ID())
+				contract.AssertNoError(err)
+				state.Provider = ref.String()
+			}
+
+			// Add to aliased maps
+			for _, alias := range state.Aliases {
+				if otherUrn, has := aliased[alias]; has {
+					contract.Assertf(!has, "two resources ('%s' and '%s') aliased to the same: '%s'", otherUrn, state.URN, alias)
+				}
+				aliased[alias] = state.URN
+			}
+		}
 	}
 }