+++ pkg/resource/deploy/source_eval.go
@@ -298,7 +298,7 @@ func (d *defaultProviders) newRegisterDefaultProviderEvent(
 	event := &registerResourceEvent{
 		goal: resource.NewGoal(
 			providers.MakeProviderType(req.Package()),
-			req.Name(), true, inputs, "", false, nil, "", nil, nil, false, nil, nil),
+			req.Name(), true, inputs, "", false, nil, "", nil, nil, false, nil, nil, nil),
 		done: done,
 	}
 	return event, done, nil
@@ -703,6 +703,11 @@ func (rm *resmon) RegisterResource(ctx context.Context,
 		provider = ref.String()
 	}
 
+	aliases := []resource.URN{}
+	for _, aliasURN := range req.GetAliases() {
+		aliases = append(aliases, resource.URN(aliasURN))
+	}
+
 	dependencies := []resource.URN{}
 	for _, dependingURN := range req.GetDependencies() {
 		dependencies = append(dependencies, resource.URN(dependingURN))
@@ -750,7 +755,7 @@ func (rm *resmon) RegisterResource(ctx context.Context,
 	// Send the goal state to the engine.
 	step := &registerResourceEvent{
 		goal: resource.NewGoal(t, name, custom, props, parent, protect, dependencies, provider, nil,
-			propertyDependencies, deleteBeforeReplace, ignoreChanges, additionalSecretOutputs),
+			propertyDependencies, deleteBeforeReplace, ignoreChanges, additionalSecretOutputs, aliases),
 		done: make(chan *RegisterResult),
 	}